requirements.txt
numpy
numba
scipy
matplotlib
pyFFTW
pyyaml
